chore: update mod structure
