+++ crates/core/src/lib.rs
@@ -1,3 +1,3 @@
-mod find_imports;
+mod core_explorer;
 
-pub use find_imports::*;
+pub use core_explorer::*;
